marketplace/requirements.txt
Flask==3.0.3
Werkzeug==3.0.1
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
click==8.1.7
flask_sqlalchemy
psycopg2-binary


